scripts/requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
pandas==2.2.2
psycopg2-binary==2.9.9
boto3==1.35.41
